move album header color methods to composables
