+++ src/composables/fetch/folders.ts
@@ -1,6 +1,6 @@
 import { Folder, Track } from "@/interfaces";
 import state from "../state";
-import useAxios from "../useAxios";
+import useAxios from "./useAxios";
 
 export default async function (path: string) {
   interface FolderData {
